refactor(drs): extract DRSSettings shape + defaults to a shared module
Sonar's duplication detector flagged a 33-line block between the
client-side DRSSettingsPanel and the server-side `/drs/settings` API
route — both maintained parallel copies of every DRSSettings field
default. New `drsSettings.ts` module is the single source of truth.

- New `frontend/src/components/automation/drs/drsSettings.ts` exports
  the DRSSettings interface, ClusterVersionInfo, and
  defaultDRSSettings. No React or MUI imports; usable from both
  client components and Next.js server routes.
- DRSSettingsPanel imports from the new module and re-exports for
  back-compat (page.tsx still does
  `import { defaultDRSSettings, type DRSSettings, type ClusterVersionInfo }
  from '@/components/automation/drs/DRSSettingsPanel'`).
- `/drs/settings` route imports defaultDRSSettings directly and aliases
  it as `defaultSettings` for the rest of its existing references.

No behavioral change. Just a single source of truth so the two files
can't drift on a default value again.
